requirements.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.11
django-debug-toolbar==1.8
django-rest-swagger==2.1.2
djangorestframework==3.7.7
gunicorn==19.4.5
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
psycopg2==2.7.3.1
pytz==2018.3
requests==2.18.4
simplejson==3.13.2
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.22
whitenoise==3.3.1
